Added solution for spy_snippits that doesn't pass google tests
For some reason I couldn’t get this to pass the last google test. I
submitted it in anguish, but it turns out since it was wrong I couldn’t
go back and I was on to level 4 with no mention of a recruiter. I
deleted my data, and I was able to get back on and I’m going to see if
I can make it back.